requirements.txt
streamlit==1.10.0
PyPDF2==2.10.3
python-docx==0.8.11
requests==2.28.1
altair==4.2.2
python-dotenv==1.0.1